+++ frontend/src/components/visual-editor/v2/AdvancedLink/AdvancedLinkModel.tsx
@@ -1,0 +1,13 @@
+import {
+  DefaultLinkModel,
+  DefaultLinkModelOptions,
+} from "@projectstorm/react-diagrams";
+
+export class AdvancedLinkModel extends DefaultLinkModel {
+  constructor(options?: DefaultLinkModelOptions) {
+    super({
+      ...options,
+      type: "advanced",
+    });
+  }
+}
