+++ frontend/src/components/visual-editor/v2/Diagrams.tsx
@@ -53,7 +53,7 @@ import { IBlock } from "@/types/block.types";
 import { ICategory } from "@/types/category.types";
 import { BlockPorts } from "@/types/visual-editor.types";
 
-import { AdvancedLinkModel } from "./AdvancedLink/AdvancedLink";
+import { AdvancedLinkModel } from "./AdvancedLink/AdvancedLinkModel";
 import BlockDialog from "../BlockDialog";
 import { ZOOM_LEVEL } from "../constants";
 import { useVisualEditor } from "../hooks/useVisualEditor";
